+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/ResourceGuiContainerHandler.java
@@ -3,8 +3,8 @@ package com.refinedmods.refinedstorage.jei.common;
 import java.util.Optional;
 import javax.annotation.Nullable;
 
-import com.refinedmods.refinedstorage2.platform.api.recipemod.IngredientConverter;
 import com.refinedmods.refinedstorage2.platform.api.support.resource.PlatformResourceKey;
+import com.refinedmods.refinedstorage2.platform.api.support.resource.RecipeModIngredientConverter;
 import com.refinedmods.refinedstorage2.platform.common.support.AbstractBaseScreen;
 import mezz.jei.api.gui.handlers.IGuiContainerHandler;
 import mezz.jei.api.ingredients.IIngredientHelper;
@@ -13,10 +13,11 @@ import mezz.jei.api.runtime.IClickableIngredient;
 import mezz.jei.api.runtime.IIngredientManager;
 
 class ResourceGuiContainerHandler implements IGuiContainerHandler<AbstractBaseScreen<?>> {
-    private final IngredientConverter converter;
+    private final RecipeModIngredientConverter converter;
     private final IIngredientManager ingredientManager;
 
-    ResourceGuiContainerHandler(final IngredientConverter converter, final IIngredientManager ingredientManager) {
+    ResourceGuiContainerHandler(final RecipeModIngredientConverter converter,
+                                final IIngredientManager ingredientManager) {
         this.converter = converter;
         this.ingredientManager = ingredientManager;
     }